Introduce a product grid and a tile partial to render products
Also add the grid to the index where data/products.json will be used to
render the products. The product grid partial will take a title and
products. A product tile will take a single product and will render a
link with an image, a title, a summary and 'read more'.